+++ app/src/main/java/com/xioneko/android/nekoanime/ui/player/AnimePlayScreen.kt
@@ -52,6 +52,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
+import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
@@ -75,6 +76,7 @@ import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.media3.common.util.UnstableApi
+import androidx.media3.exoplayer.ExoPlayer
 import androidx.media3.exoplayer.offline.Download.STATE_COMPLETED
 import androidx.media3.exoplayer.offline.Download.STATE_DOWNLOADING
 import androidx.media3.exoplayer.offline.Download.STATE_QUEUED
@@ -84,9 +86,14 @@ import com.xioneko.android.nekoanime.data.AnimeDownloadHelper.Companion.STATE_PR
 import com.xioneko.android.nekoanime.data.model.Anime
 import com.xioneko.android.nekoanime.data.model.AnimeShell
 import com.xioneko.android.nekoanime.data.model.asAnimeShell
+import com.xioneko.android.nekoanime.data.network.danmu.api.DanmuSession
+import com.xioneko.android.nekoanime.data.network.danmu.api.DanmukuEvent
+import com.xioneko.android.nekoanime.data.network.danmu.dto.DanmakuPresentation
 import com.xioneko.android.nekoanime.ui.component.AnimatedFollowIcon
 import com.xioneko.android.nekoanime.ui.component.BottomSheet
 import com.xioneko.android.nekoanime.ui.component.LazyAnimeGrid
+import com.xioneko.android.nekoanime.ui.danmu.DanmakuConfigData
+import com.xioneko.android.nekoanime.ui.danmu.rememberDanmakuHostState
 import com.xioneko.android.nekoanime.ui.theme.NekoAnimeIcons
 import com.xioneko.android.nekoanime.ui.theme.basicBlack
 import com.xioneko.android.nekoanime.ui.theme.basicWhite
@@ -100,13 +107,16 @@ import com.xioneko.android.nekoanime.ui.theme.pink50
 import com.xioneko.android.nekoanime.ui.theme.pink60
 import com.xioneko.android.nekoanime.ui.theme.pink70
 import com.xioneko.android.nekoanime.ui.theme.pink97
+import com.xioneko.android.nekoanime.ui.util.KEY_DANMAKU_CONFIG_DATA
 import com.xioneko.android.nekoanime.ui.util.LoadingState
 import com.xioneko.android.nekoanime.ui.util.currentScreenSizeDp
 import com.xioneko.android.nekoanime.ui.util.isOrientationLocked
 import com.xioneko.android.nekoanime.ui.util.isTablet
+import com.xioneko.android.nekoanime.ui.util.rememberPreference
 import com.xioneko.android.nekoanime.ui.util.setScreenOrientation
 import kotlinx.coroutines.delay
 import kotlin.math.max
+import kotlin.time.Duration.Companion.milliseconds
 
 @ExperimentalLayoutApi
 @ExperimentalMaterial3Api
@@ -915,6 +925,57 @@ private fun OfflineCacheEpisodesGrid(
                 }
             }
 
+        }
+    }
+}
+
+@Composable
+fun DanmakuHost(
+    playerState: NekoAnimePlayerState,
+    session: DanmuSession?,
+    enabled: Boolean,
+    player: ExoPlayer
+) {
+    if (!enabled) return
+    val danmakuConfigData by rememberPreference(
+        KEY_DANMAKU_CONFIG_DATA,
+        DanmakuConfigData(),
+        DanmakuConfigData.serializer()
+    )
+    val danmakuHostState =
+        rememberDanmakuHostState(danmakuConfig = danmakuConfigData.toDanmakuConfig())
+
+    if (session != null) {
+        com.xioneko.android.nekoanime.ui.danmu.DanmakuHost(state = danmakuHostState)
+    }
+
+    LaunchedEffect(playerState.isPlaying) {
+        if (playerState.isPlaying) {
+            danmakuHostState.play()
+        } else {
+            danmakuHostState.pause()
+        }
+    }
+
+    val isPlayingFlow = remember { snapshotFlow { player.isPlaying } }
+    LaunchedEffect(session) {
+        danmakuHostState.clearPresentDanmaku()
+        session?.at(
+            curTimeMillis = { player.currentPosition.milliseconds },
+            isPlayingFlow = isPlayingFlow,
+        )?.collect { danmakuEvent ->
+            when (danmakuEvent) {
+                is DanmukuEvent.Add -> {
+                    danmakuHostState.trySend(
+                        DanmakuPresentation(
+                            danmakuEvent.danmu,
+                            false
+                        )
+                    )
+                }
+                // 快进/快退
+                is DanmukuEvent.Repopulate -> danmakuHostState.repopulate()
+            }
         }
     }
 }
